+++ src/Pages/Main/ArticleNewProduct/ArticleNewProduct.scss
@@ -5,7 +5,7 @@
 
   .articleNewProductBox {
     height: 645px;
-    padding: 82px 0px 50px 0px;
+    padding: 60px 0px 50px 0px;
     background-color: #f7f7f7;
 
     .newProductTitle {
