+++ src/Pages/Main/Banner/Banner.scss
@@ -2,6 +2,7 @@
 
 .banner {
   position: relative;
+  height: 400px;
 
   .leftButton {
     position: absolute;
