+++ src/Pages/Nav/Nav.scss
@@ -1,6 +1,6 @@
 @import '../../Style/common.scss';
+
 .nav {
-  border-bottom: 2px solid #f7f7f7;
   font-family: $fontFamily;
 
   .navBox {
@@ -104,59 +104,6 @@
         color: #fff;
         background-color: $orangeColor;
         border-radius: 50%;
-      }
-    }
-  }
-
-  .navCategory {
-    @extend %flexCenter;
-    width: 1280px;
-    height: 64px;
-    margin: 0 auto;
-    font-family: $fontFamily;
-
-    .categoryLeft {
-      width: 935px;
-      height: 26px;
-
-      .categoryLeftUl {
-        .fa-bars {
-          margin-right: 10px;
-          color: #333;
-          font-size: 20px;
-        }
-
-        .categoryLeftLi {
-          display: inline-block;
-          margin-right: 28px;
-          color: #333;
-
-          .categoryLeftLink {
-            color: #333333;
-            text-decoration: none;
-          }
-        }
-      }
-    }
-
-    .categoryRight {
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      width: 346px;
-      height: 26px;
-
-      .categoryRightUl {
-        .categoryRightLi {
-          display: inline-block;
-          margin-left: 35px;
-          color: #333;
-
-          .categoryRightLink {
-            color: #333333;
-            text-decoration: none;
-          }
-        }
       }
     }
   }
