+++ src/Pages/Nav/NavCategory.scss
@@ -1,0 +1,57 @@
+@import '../../Style/common.scss';
+
+.navCategory {
+  @extend %flexCenter;
+  position: sticky;
+  height: 64px;
+  margin: 0 auto;
+  background-color: #fff;
+  font-family: $fontFamily;
+  top: 0;
+  z-index: 1;
+
+  .categoryLeft {
+    width: 935px;
+    height: 26px;
+
+    .categoryLeftUl {
+      .fa-bars {
+        margin-right: 10px;
+        color: #333;
+        font-size: 20px;
+      }
+
+      .categoryLeftLi {
+        display: inline-block;
+        margin-right: 28px;
+        color: #333;
+
+        .categoryLeftLink {
+          color: #333333;
+          text-decoration: none;
+        }
+      }
+    }
+  }
+
+  .categoryRight {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    width: 346px;
+    height: 26px;
+
+    .categoryRightUl {
+      .categoryRightLi {
+        display: inline-block;
+        margin-left: 35px;
+        color: #333;
+
+        .categoryRightLink {
+          color: #333333;
+          text-decoration: none;
+        }
+      }
+    }
+  }
+}
